% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_basics.R
\name{geom_hypo_LG}
\alias{geom_hypo_LG}
\title{Shaded LG backgrounds}
\usage{
geom_hypo_LG(...)
}
\arguments{
\item{...}{catch all parameter (legacy)}
}
\description{
\code{geom_hypo_LG} adds the linkage group (LG) boundaries to background of a ggplot.
}
\details{
To add context to genome wide plots of a statistic (e.g. Fst) this function
adds a alternating shaded background to the plot.
}
\examples{
ggplot() +
  geom_hypo_LG() +
  scale_fill_hypo_LG_bg() +
  scale_x_hypo_LG() +
  coord_cartesian(ylim = c(0, 1)) +
  theme_hypo()

}
\seealso{
\code{\link{scale_fill_hypo_LG_bg}}
}
