% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_geom_hypobg.R
\name{geom_hypobg}
\alias{geom_hypobg}
\title{Add linkage group annotation}
\usage{
geom_hypobg(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes()}

\item{data}{The data to be displayed in this layer - usually hypogen::hypo_karyotype}

\item{stat}{string, "identity"}

\item{position}{Position adjustment "identity"}

\item{...}{Other arguments passed on to layer().}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{logical}
}
\description{
Add linkage group annotation
}
