% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_geom_hypobg.R
\name{guide_colorbar_hypobg}
\alias{guide_colorbar_hypobg}
\title{Clone of guide_colorbar for linkage group background}
\usage{
guide_colorbar_hypobg(
  title = waiver(),
  title.position = NULL,
  title.theme = NULL,
  title.hjust = NULL,
  title.vjust = NULL,
  label = TRUE,
  label.position = NULL,
  label.theme = NULL,
  label.hjust = NULL,
  label.vjust = NULL,
  barwidth = NULL,
  barheight = NULL,
  nbin = 20,
  raster = TRUE,
  frame.colour = NULL,
  frame.linewidth = 0.5,
  frame.linetype = 1,
  ticks = TRUE,
  ticks.colour = "white",
  ticks.linewidth = 0.5,
  draw.ulim = TRUE,
  draw.llim = TRUE,
  direction = NULL,
  default.unit = "line",
  reverse = FALSE,
  order = 0,
  available_aes = c("colour", "color", "fill", "hypobg"),
  ...
)
}
\arguments{
\item{title}{clone, s. ggplot2::guide_colorbar()}

\item{title.position}{clone, s. ggplot2::guide_colorbar()}

\item{title.theme}{clone, s. ggplot2::guide_colorbar()}

\item{title.hjust}{clone, s. ggplot2::guide_colorbar()}

\item{title.vjust}{clone, s. ggplot2::guide_colorbar()}

\item{label}{clone, s. ggplot2::guide_colorbar()}

\item{label.position}{clone, s. ggplot2::guide_colorbar()}

\item{label.theme}{clone, s. ggplot2::guide_colorbar()}

\item{label.hjust}{clone, s. ggplot2::guide_colorbar()}

\item{label.vjust}{clone, s. ggplot2::guide_colorbar()}

\item{barwidth}{clone, s. ggplot2::guide_colorbar()}

\item{barheight}{clone, s. ggplot2::guide_colorbar()}

\item{nbin}{clone, s. ggplot2::guide_colorbar()}

\item{raster}{clone, s. ggplot2::guide_colorbar()}

\item{frame.colour}{clone, s. ggplot2::guide_colorbar()}

\item{frame.linewidth}{clone, s. ggplot2::guide_colorbar()}

\item{frame.linetype}{clone, s. ggplot2::guide_colorbar()}

\item{ticks}{clone, s. ggplot2::guide_colorbar()}

\item{ticks.colour}{clone, s. ggplot2::guide_colorbar()}

\item{ticks.linewidth}{clone, s. ggplot2::guide_colorbar()}

\item{draw.ulim}{clone, s. ggplot2::guide_colorbar()}

\item{draw.llim}{clone, s. ggplot2::guide_colorbar()}

\item{direction}{clone, s. ggplot2::guide_colorbar()}

\item{default.unit}{clone, s. ggplot2::guide_colorbar()}

\item{reverse}{clone, s. ggplot2::guide_colorbar()}

\item{order}{clone, s. ggplot2::guide_colorbar()}

\item{available_aes}{string, extended ggplot2::guide_colorbar() options by "hypobg"}

\item{...}{clone, s. ggplot2::guide_colorbar()}
}
\description{
Clone of guide_colorbar for linkage group background
}
