% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_data.R
\docType{data}
\name{hypo_chrom_start}
\alias{hypo_chrom_start}
\title{Hamlet Genome Linkage Group Start Positions.}
\format{
A tibble with 24 rows and 2 variables:
\describe{
\item{CHROM}{string, linkage group ID, eg ("LG01")}
\item{GSTART}{integer, genomic start positions of the linkage groups (bp)}
}
}
\source{
\url{https://doi.org/10.1038/s41559-019-0814-5}
}
\usage{
hypo_chrom_start
}
\description{
A dataset containing the genomic start positions of the linkage groups of the Hamlet reference genome.
}
\keyword{datasets}
