% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_data.R
\docType{data}
\name{hypo_clr_lg}
\alias{hypo_clr_lg}
\title{Hamlet Genome LG Background Color.}
\format{
An object of class \code{character} of length 1.
}
\usage{
hypo_clr_lg
}
\description{
A default background color for "even" linkage groups of the hamlet genome.
}
\keyword{datasets}
