% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_data.R
\docType{data}
\name{hypo_clr_LGs}
\alias{hypo_clr_LGs}
\title{Hamlet Genome Color Scheme.}
\format{
An object of class \code{character} of length 26.
}
\usage{
hypo_clr_LGs
}
\description{
A color scheme for the hamlet genome consisting of 24 colors for the 24 linkage groups.
}
\keyword{datasets}
