% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_import.R
\name{hypo_hwe}
\alias{hypo_hwe}
\title{Compute Hardy-Weinberg genopye frequencies}
\usage{
hypo_hwe(n = 100)
}
\arguments{
\item{n}{integer skalar (>= 2, mandatory), length of the
allele freqency range}
}
\description{
\code{hypo_hwe} computes Hardy-Weinberg genopye frequencies for p ranging from 0 to 1.
}
\details{
This function creates a range of allele frequencies ranging from p = 0 to p = 1
and computes the respective genotype frequencies according to Hardy-Weinberg:

\itemize{
  \item{\eqn{q = 1 - p}}
  \item{\eqn{AA = p^2}}
  \item{\eqn{Aa = 2pq}}
  \item{\eqn{aa = q^2}}
}

The function returns the results as table.
}
\examples{

genotype_hwe <- hypo_hwe(100)

}
