% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_import.R
\name{hypo_import_snps}
\alias{hypo_import_snps}
\title{Import SNP based statistics}
\usage{
hypo_import_snps(file, gz = FALSE, run, ...)
}
\arguments{
\item{file}{string scalar (mandatory), the input file}

\item{gz}{logical scalar (optional), is the input file gz compressed?}

\item{run}{string scalar (optional), appends a column RUN as ID if several files should be merged later}

\item{...}{parameters passed to vroom::vroom()}
}
\description{
\code{hypo_import_snps} imports SNP based statistics (such as VCFtools output).
}
\details{
This function imports SNP based statistics. The input is
expected to be tab separated, to contain a CHROM and POS column
and can optionally be gz compressed.
After import, the CHROM based position is transposed to a genome
wide position for continuous visualization.
}
\examples{

file_snps <- system.file("extdata", "example.weir.fst.gz", package = "hypogen")

hypo_import_snps(file = file_snps, gz = TRUE)

}
\seealso{
\code{\link{hypo_import_windows}}
}
