% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_data.R
\docType{data}
\name{hypo_karyotype}
\alias{hypo_karyotype}
\title{Hamlet Genome Karyotype.}
\format{
A tibble with 24 rows and 5 variables:
\describe{
\item{CHROM}{string, linkage group ID, eg ("LG01")}
\item{LENGTH}{integer, length of the linkage groups (bp)}
\item{GSTART}{integer, genomic start positions of the linkage groups (bp)}
\item{GEND}{integer, genomic end positions of the linkage groups (bp)}
\item{GROUP}{string ("odd"/"even"), alternating across linagke groups}
}
}
\source{
\url{https://doi.org/10.1038/s41559-019-0814-5}
}
\usage{
hypo_karyotype
}
\description{
A dataset containing the hamlet karyotype according to the first reference genome.
}
\keyword{datasets}
