% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_basics.R
\name{hypo_which_CHROM}
\alias{hypo_which_CHROM}
\title{Determining the LG from the genomic positions}
\usage{
hypo_which_CHROM(x)
}
\arguments{
\item{x}{interger vector (mandatory), a vector of bp positions on the hamlet genome.}
}
\description{
\code{hypo_which_CHROM} determines the LG from a vector of genomic position.
}
\details{
The function vectorizes the \code{hypo_which_CHROM_s} function.
}
\seealso{
\code{\link{hypo_which_CHROM}}
}
