% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_basics.R
\docType{package}
\name{hypogen}
\alias{hypogen}
\title{hypogen: Provides Hypoplectrus PopGen Data and Functions}
\description{
Basic reference data for population genomic studies in Caribbean
hamlets (Hypoplectrus sp.). It includes summary data of the Hamlet reference
genome (accession number: PRJEB27858, Hench et al. 2018) as well its annotation.
Furthermore it includes some functions to use the reference data  to put new data
into context.
}
