% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_basics.R
\name{scale_color_hypo_LG}
\alias{scale_color_hypo_LG}
\title{Set LG color palette}
\usage{
scale_color_hypo_LG(..., name = "LG")
}
\description{
\code{scale_color_hypo_LG} adjusts the color scale with default color for each LG.
}
\details{
Each of the LGs is assigned to one default color.

\strong{! CAREFUL:} should not be used if additional color levels exiest.
}
\seealso{
\code{\link{geom_hypo_LG}}, \code{\link{scale_x_hypo_LG}}
}
