% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_basics.R
\name{scale_fill_hypo_LG_bg}
\alias{scale_fill_hypo_LG_bg}
\title{Set LG fill}
\usage{
scale_fill_hypo_LG_bg(..., values = c(odd = NA, even = hypo_clr_lg))
}
\description{
\code{scale_fill_hypo_LG_bg} sets the color scheme for the LG background shading.
}
\details{
Fill values are set for tnhe two LG classes (even/odd LG nr).

\strong{! CAREFUL:} LG fill values are set to "zz1" $ zz2.
If additional fill levels exist, the fill values of the LGs should be appended
to the desired fill scale.
}
\seealso{
\code{\link{geom_hypo_LG}}, \code{\link{scale_x_hypo_LG}}, \code{\link{scale_color_hypo_LG}}
}
