% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_geom_hypobg.R
\name{scale_hypobg_distiller}
\alias{scale_hypobg_distiller}
\title{Scale Linkage Group background}
\usage{
scale_hypobg_distiller(
  ...,
  type = "seq",
  palette = 1,
  direction = -1,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "hypobg"
)
}
\arguments{
\item{...}{parameters passed to ggplot2::continuous_scale()}

\item{type}{string, RColorBrewer palette type}

\item{palette}{integer, RColorBrewer palette index}

\item{direction}{numeric (1/ -1)}

\item{values}{numeric vector, passed to scales::gradient_n_pal}

\item{space}{colour space in which to calculate gradient, passed to scales::gradient_n_pal}

\item{na.value}{Missing values will be replaced with this value.}

\item{guide}{ggplot2::guide}

\item{aesthetics}{string ("hypobg")}
}
\description{
Scale Linkage Group background
}
