% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_geom_hypobg.R
\name{scale_hypobg_manual}
\alias{scale_hypobg_manual}
\title{Manual linkage group color scale}
\usage{
scale_hypobg_manual(..., values, aesthetics = "hypobg")
}
\arguments{
\item{...}{parameters passed to ggplot2:::manual_scale()}

\item{values}{string vector, color values}

\item{aesthetics}{string, "hypobg"}
}
\description{
Manual linkage group color scale
}
