% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_basics.R
\name{scale_x_hypo_LG}
\alias{scale_x_hypo_LG}
\title{Format genome wide x axis}
\usage{
scale_x_hypo_LG(..., name = "", expand = c(0, 0),
  breaks = (hypo_karyotype$GSTART + hypo_karyotype$GEND)/2, labels = 1:24,
  position = "top")
}
\description{
\code{scale_x_hypo_LG} adjusts the x axis to match the hamet genome.
}
\details{
Each of the LGs is labelled with on tick which is placed in the
center of the LG.
The should be combined with \code{hypogen::geom_hypo_LG}
}
\examples{

ggplot() +
  geom_hypo_LG() +
  scale_x_hypo_LG()
}
\seealso{
\code{\link{geom_hypo_LG}}, \code{\link{scale_color_hypo_LG}}
}
