% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_basics.R
\name{theme_hypo}
\alias{theme_hypo}
\title{Set hypogen plot theme}
\usage{
theme_hypo(...)
}
\arguments{
\item{...}{parameters passed to ggplot2::theme()}
}
\description{
\code{theme_hypo} sets the default hypogen plot theme.
}
\details{
This theme is optimized for genome wide plots.
}
\seealso{
\code{\link{theme_hypo_anno_extra}}
}
