% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_basics.R
\name{theme_hypo_anno_extra}
\alias{theme_hypo_anno_extra}
\title{Set hypogen plot theme}
\usage{
theme_hypo_anno_extra(...)
}
\arguments{
\item{...}{catch all parameter (legacy)}
}
\description{
\code{theme_hypo_anno_extra} adjusts the theme for annotation plots.
}
\seealso{
\code{\link{theme_hypo}}
}
