% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_basics.R
\name{geom_hypo_LG}
\alias{geom_hypo_LG}
\title{Shaded LG backgrounds}
\usage{
geom_hypo_LG(...)
}
\description{
\code{geom_hypo_LG} adds the linkage group (LG) boundaries to background of a ggplot.
}
\details{
To add context to genome wide plots of a statistic (e.g. Fst) this function
adds a alternating shaded background to the plot.
}
\examples{

}
\seealso{
\code{\link{scale_fill_hypo_LG_bg}}

ggplot() +
  geom_hypo_LG()+
  scale_color_hypo_LG()
}
