% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_import.R
\name{hypo_import_snps}
\alias{hypo_import_snps}
\title{Import SNP based statistics}
\usage{
hypo_import_snps(file, gz = FALSE, ...)
}
\arguments{
\item{file}{string skalar (mandatory), the input file}

\item{gz}{logical skalar (optional), is the input file gz compressed?}
}
\description{
\code{hypo_import_snps} imports SNP based statistics (such as VCFtools output).
}
\details{
This function imports SNP based statistics. The input is
expected to be tab separated, to contain a CHROM and POS column
and can optionally be gz compressed.
After import, the CHROM based position is transposed to a genome
wide position for continuous visualization.
}
\examples{

file_snps <- system.file("extdata", "example.weir.fst.gz", package = "hypogen")

hypo_import_snps(file = file_snps, gz = TRUE)

}
\seealso{
\code{\link{hypo_import_windows}}
}
