% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypogen_basics.R
\name{hypo_which_CHROM_s}
\alias{hypo_which_CHROM_s}
\title{Determining the LG from the genomic position}
\usage{
hypo_which_CHROM_s(x)
}
\arguments{
\item{x}{integer scalar (mandatory), a bp position on the hamlet genome.}
}
\description{
\code{hypo_which_CHROM_s} determines the LG from the genomic position.
}
\details{
The function finds the respective LG on which a specific genomic position (1 - 559,649,677)
resides.
}
\examples{
hypo_which_CHROM_s(123581321)

}
\seealso{
\code{\link{hypo_which_CHROM}}
}
